+++ templates/footer.php
@@ -1,10 +1,62 @@
 
-<footer>
-    <div class="container">
-        Copyright WWI
-    </div>
-</footer>
+<!-- Footer -->
+<footer class="bg-dark text-center text-white">
+    <!-- Grid container -->
+    <div class="container p-4">
+        <!-- Section: Social media -->
+        <section class="mb-4">
+            <!-- Facebook -->
+            <a class="btn btn-outline-light btn-floating m-1" href="https://github.com/tromphakvoort/WorldWideImporters" role="button"
+            ><i class="fab fa-facebook-f"></i
+                ></a>
 
+            <!-- Twitter -->
+            <a class="btn btn-outline-light btn-floating m-1" href="https://github.com/tromphakvoort/WorldWideImporters" role="button"
+            ><i class="fab fa-twitter"></i
+                ></a>
+
+            <!-- Google -->
+            <a class="btn btn-outline-light btn-floating m-1" href="https://github.com/tromphakvoort/WorldWideImporters" role="button"
+            ><i class="fab fa-google"></i
+                ></a>
+
+            <!-- Instagram -->
+            <a class="btn btn-outline-light btn-floating m-1" href="https://github.com/tromphakvoort/WorldWideImporters" role="button"
+            ><i class="fab fa-instagram"></i
+                ></a>
+
+            <!-- Linkedin -->
+            <a class="btn btn-outline-light btn-floating m-1" href="https://github.com/tromphakvoort/WorldWideImporters" role="button"
+            ><i class="fab fa-linkedin-in"></i
+                ></a>
+
+            <!-- Github -->
+            <a class="btn btn-outline-light btn-floating m-1" href="https://github.com/tromphakvoort/WorldWideImporters" role="button"
+            ><i class="fab fa-github"></i
+                ></a>
+        </section>
+        <!-- Section: Social media -->
+
+        <!-- Section: Text -->
+        <section class="mb-4">
+            <p>
+                Een heerlijke gekoelde chocoladereep gemaakt van de beste cacaobonen uit Peru geselecteerd door onze meesterchocolatiers. <br />
+                Curabitur blandit tempus porttitor. Donec ullamcorper nulla non metus auctor fringilla. Cras mattis consectetur purus sit amet fermentum.
+            </p>
+        </section>
+        <!-- Section: Text -->
+    </div>
+    <!-- Grid container -->
+
+    <!-- Copyright -->
+    <div class="text-center p-3" style="background-color: rgba(0, 0, 0, 0.2);">
+        © 2022 Copyright:
+        <a class="text-white" href="https://github.com/tromphakvoort/WorldWideImporters">World Wide Importers</a>
+    </div>
+    <!-- Copyright -->
+</footer>
+<!-- Footer -->
+<script src="https://kit.fontawesome.com/7e21b455c6.js" crossorigin="anonymous"></script>
 <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.10.2/dist/umd/popper.min.js" integrity="sha384-7+zCNj/IqJ95wo16oMtfsKbZ9ccEh31eOz1HGyDuCQ6wgnyJNSYdrPa03rtR1zdB" crossorigin="anonymous"></script>
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.min.js" integrity="sha384-QJHtvGhmr9XOIpI6YVutG+2QOK9T+ZnN4kzFN1RtK3zEFEIsxhlmWl5/YESvpZ13" crossorigin="anonymous"></script>
 </body>