+++ views/cart.php
@@ -31,8 +31,7 @@ $results = mysqli_query($connection, "SELECT * FROM products ORDER BY id ASC");
 
 
 ?>
-    <div class="container" style="width: 65%">
-        <div style="clear: both"></div>
+    <div class="container">
         <h3 class="title2">Winkelwagen</h3>
         <div class="table-responsive">
             <table class="table table-bordered">
@@ -52,11 +51,16 @@ $results = mysqli_query($connection, "SELECT * FROM products ORDER BY id ASC");
                     foreach ($_SESSION["cart"] as $product) {
                         echo "
                 <tr>
-                    <td>" . $product->getProductName    () . "</td>
+                    <td>" . $product->getProductName() . "</td>
                     <td>1</td>
-                    <td>" . round($product->getPriceAmount() / 100, $product->getPricePrecision()) . "</td>
-                    <td>" . round($product->getPriceAmount() / 100, $product->getPricePrecision()) * 1 . "</td>
-                    <td></td>
+                    <td>€ " . round($product->getPriceAmount() / 100, $product->getPricePrecision()) . "</td>
+                    <td>€ " . round($product->getPriceAmount() / 100, $product->getPricePrecision()) * 1 . "</td>
+                    <td>
+                        <form action='/removeFromCart' method='post'>
+                            <input type='hidden' name='product-id' value='" . $product->getId() . "' />
+                            <button class='btn btn-danger' type='submit'>Verwijderen</button>
+                        </form>
+                    </td>
                 </tr> ";
 
                     $total = $total + (round($product->getPriceAmount() / 100, $product->getPricePrecision()) * 1   );
@@ -64,12 +68,13 @@ $results = mysqli_query($connection, "SELECT * FROM products ORDER BY id ASC");
                 echo "
                 <tr>
                     <td colspan='3' align='right'>Totaal</td>
-                    <th align='right'>$ " . number_format($total, 2) . "</th>
+                    <th align='right'>€ " . number_format($total, 2) . "</th>
                     <td></td>
                 </tr>";
                     }; ?>
             </table>
         </div>
+        <a href="/placeOrder" class="btn btn-warning btn-lg">Afrekenen</a>
     </div>