+++ views/placeOrder.php
@@ -1,0 +1,100 @@
+<?php $title = "Homepage | World Wide Importers";
+// require_once("CSS/Style.css");
+include("../templates/header.php");
+
+
+?>
+
+<div class="container">
+    <div class="row">
+        <div class="col-md-12 mb-5">
+            <h1>Afrekenen - PR28349525</h1>
+        </div>
+
+        <div class="col-md-3">
+            <div class="card" aria-hidden="true">
+                <img src="https://res.cloudinary.com/practicaldev/image/fetch/s--ExVB4I1A--/c_limit%2Cf_auto%2Cfl_progressive%2Cq_66%2Cw_880/https://dev-to-uploads.s3.amazonaws.com/i/mrvsmk2pl3l8fwocbfhy.gif"
+                     class="card-img-top" alt="Plcaeholder">
+                <div class="card-body">
+                    <h5 class="card-title placeholder-glow">
+                        <span class="placeholder col-6"></span>
+                    </h5>
+                    <p class="card-text placeholder-glow">
+                        <span class="placeholder col-7"></span>
+                        <span class="placeholder col-4"></span>
+                        <span class="placeholder col-4"></span>
+                        <span class="placeholder col-6"></span>
+                        <span class="placeholder col-8"></span>
+                    </p>
+                    <a href="#" tabindex="-1" class="btn btn-primary disabled placeholder col-6"></a>
+                </div>
+            </div>
+        </div>
+
+        <div class="col-md-3">
+            <div class="card" aria-hidden="true">
+                <img src="https://res.cloudinary.com/practicaldev/image/fetch/s--ExVB4I1A--/c_limit%2Cf_auto%2Cfl_progressive%2Cq_66%2Cw_880/https://dev-to-uploads.s3.amazonaws.com/i/mrvsmk2pl3l8fwocbfhy.gif"
+                     class="card-img-top" alt="Plcaeholder">
+                <div class="card-body">
+                    <h5 class="card-title placeholder-glow">
+                        <span class="placeholder col-6"></span>
+                    </h5>
+                    <p class="card-text placeholder-glow">
+                        <span class="placeholder col-7"></span>
+                        <span class="placeholder col-4"></span>
+                        <span class="placeholder col-4"></span>
+                        <span class="placeholder col-6"></span>
+                        <span class="placeholder col-8"></span>
+                    </p>
+                    <a href="#" tabindex="-1" class="btn btn-info disabled placeholder col-6"></a>
+                </div>
+            </div>
+        </div>
+
+        <div class="col-md-3">
+            <div class="card" aria-hidden="true">
+                <img src="https://res.cloudinary.com/practicaldev/image/fetch/s--ExVB4I1A--/c_limit%2Cf_auto%2Cfl_progressive%2Cq_66%2Cw_880/https://dev-to-uploads.s3.amazonaws.com/i/mrvsmk2pl3l8fwocbfhy.gif"
+                     class="card-img-top" alt="Plcaeholder">
+                <div class="card-body">
+                    <h5 class="card-title placeholder-glow">
+                        <span class="placeholder col-6"></span>
+                    </h5>
+                    <p class="card-text placeholder-glow">
+                        <span class="placeholder col-7"></span>
+                        <span class="placeholder col-4"></span>
+                        <span class="placeholder col-4"></span>
+                        <span class="placeholder col-6"></span>
+                        <span class="placeholder col-8"></span>
+                    </p>
+                    <a href="#" tabindex="-1" class="btn btn-warning disabled placeholder col-6"></a>
+                </div>
+            </div>
+        </div>
+
+        <div class="col-md-3">
+            <div class="card" aria-hidden="true">
+                <img src="https://res.cloudinary.com/practicaldev/image/fetch/s--ExVB4I1A--/c_limit%2Cf_auto%2Cfl_progressive%2Cq_66%2Cw_880/https://dev-to-uploads.s3.amazonaws.com/i/mrvsmk2pl3l8fwocbfhy.gif"
+                     class="card-img-top" alt="Plcaeholder">
+                <div class="card-body">
+                    <h5 class="card-title placeholder-glow">
+                        <span class="placeholder col-6"></span>
+                    </h5>
+                    <p class="card-text placeholder-glow">
+                        <span class="placeholder col-7"></span>
+                        <span class="placeholder col-4"></span>
+                        <span class="placeholder col-4"></span>
+                        <span class="placeholder col-6"></span>
+                        <span class="placeholder col-8"></span>
+                    </p>
+                    <a href="#" tabindex="-1" class="btn btn-danger disabled placeholder col-6"></a>
+                </div>
+            </div>
+        </div>
+
+        <div class="col-md-12 mt-3">
+            <p class="small-test">Note: Dit is een dummy pagina 🤖</p>
+        </div>
+    </div>
+</div>
+
+<?php include("../templates/footer.php"); ?>